+++ llamaamp/settings.py
@@ -144,6 +144,7 @@ SCHEMA = {
     'windowshade': Setting(False, _is_true),
     'expanded_size': Setting([WINDOW_W, WINDOW_H], _window_size),
     'panels': Setting({}, _dict),
+    'vis_mode': Setting('spectrum', _choice(('spectrum', 'scope'))),
 }